Adding CI workflow
Added workflow so that test file runs on push / pull. My code is mainly from the starter workflows repo and the workflow file from the nodebb repo. I also don't know if coveralls works with this, but I copied over the coverall stuff from the other yaml file to see if it will.